backend/pyproject.toml
[project]
name = "ai_trace_datasets"
version = "0.0.1"
requires-python = ">=3.11.0"
description = "AI Trace Datasets"
dependencies = [
    "anthropic>=0.50.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "cloud-sql-python-connector[asyncpg]>=1.18.1",
    "fastapi>=0.115.12",
    "gcloud-aio-storage>=9.1.0",
    "google>=3.0.0",
    "greenlet>=3.2.1",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "ruff>=0.11.9",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py311"
